Initial Application of Erik Quanstrom's APE updates

+++ sys/src/ape/lib/bsd/gethostname.c
@@ -6,7 +6,7 @@
 #include <string.h>
 
 int
-gethostname(char *name, int namelen)
+gethostname(char *name, size_t namelen)
 {
 	int n, fd;
 	char buf[128];
